+++ lib/data/wishlist.dart
@@ -4,6 +4,7 @@ class Wishlist {
   String uuid;
   String label;
   String itemCount;
+  int color;
   Timestamp timestamp;
 
   Wishlist.fromMap(String uuid, Map<String, dynamic> properties) {
@@ -11,5 +12,6 @@ class Wishlist {
     this.label = properties["label"];
     this.itemCount = properties["itemCounts"];
     this.timestamp = properties["timestamp"];
+    this.color = properties["color"];
   }
 }