+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add settings page
 - Add recipes
 - Add possiblity to add a recipe to a list
+- Add choose wishlist color
 
 ### Changed